[Bronze V] Title: 이상한 기호, Time: 208 ms, Memory: 17688 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -3,13 +3,13 @@ import java.util.Scanner;
 class Main{
     public static void main(String[] args){
         Scanner in = new Scanner(System.in);
-        int A = in.nextInt();
-        int B = in.nextInt();
+        long A = in.nextInt();
+        long B = in.nextInt();
         System.out.print(cals(A, B));
     }
     
-    private static int cals(int a, int b){
-        int result = (a + b) * (a - b);
+    private static long cals(long a, long b){
+        long result = (a + b) * (a - b);
         return result;
     }
 }
